Cloud Function/requirements.txt
functions-framework==3.*
requests==2.27.1
jwt==1.3.1
google-auth==2.6.0
google-cloud-secret-manager==2.10.0
google-cloud-logging==3.5.0